benchmarks: fix SPSC threading and validate 74ns cross-thread pipelining
- Fixed writeCache false-sharing between Producer/Consumer in Channel
- Restored correctness validation matching the classic array FSM
- Updated latency metrics in README.md based on new concurrently unrolled topology

+++ autumn-core/src/commonMain/kotlin/dev/autumn/memory/AutumnMemoryBank.kt
@@ -10,7 +10,7 @@ expect object AutumnMemoryBank {
      * static bounds calculated by the Compiler Plugin during parsing.
      * Guaranteed to perfectly align to hardware page sizes (4096 boundaries bounds).
      */
-    fun allocate(sizeBytes: Int)
+    fun allocate(sizeBytes: Long)
     
     fun getInt(offset: Int): Int
     fun setInt(offset: Int, value: Int)
